Split concrete text for (not) being logged in up in separate components
