+++ recipe-front-end/src/components/account/AccountMenu.tsx
@@ -4,7 +4,8 @@ import { connect } from "react-redux";
 import { Localisation } from "../../localisation/AppTexts";
 import { toggleLoginForm } from "../../redux/Actions";
 import { ReduxModel } from "../../redux/Store";
-import { makeQueryString } from "../../utils/UrlUtils";
+import { LoggedInText } from "./LoggedInText";
+import { NotLoggedIn } from "./NotLoggedInText";
 
 interface AccountMenuProps {
     loginMenuOpened: boolean;
@@ -27,11 +28,6 @@ function mapStateToProps(reduxStore: ReduxModel): AccountMenuProps {
 }
 
 function AccountMenu(props: Props) {
-    const queryParams = makeQueryString({
-        client_id: process.env.REACT_APP_GITHUB_CLIENT_ID,
-        scope: 'user:email'
-    });
-
     return (<Drawer
         isOpen={props.loginMenuOpened}
         placement="right"
@@ -42,10 +38,7 @@ function AccountMenu(props: Props) {
                 <DrawerCloseButton />
                 <DrawerHeader>{Localisation.WELCOME} <b>{props.loggedIn ? props.userName : Localisation.NINJA}</b> 🐱‍👤</DrawerHeader>
                 <DrawerBody>
-                    <p>{Localisation.NOT_YOU}</p>
-                    <a href={`https://github.com/login/oauth/authorize?${queryParams}`}>
-                        {Localisation.LOGIN_FOR_MORE_FEATURES} <img src='images/github.png' />
-                    </a>
+                    {props.loggedIn ? <LoggedInText /> : <NotLoggedIn />}
                 </DrawerBody>
             </DrawerContent>
         </DrawerOverlay>
