+++ recipe-front-end/src/components/account/LoggedInText.tsx
@@ -1,0 +1,9 @@
+import { Box } from "@chakra-ui/react";
+import React from "react";
+import { Localisation } from "../../localisation/AppTexts";
+
+export function LoggedInText() {
+    return (<Box>
+        {Localisation.LOGGED_IN_ENCOURAGING}
+    </Box>)
+}
