+++ recipe-front-end/src/localisation/AppTexts.ts
@@ -7,5 +7,6 @@ export enum Localisation {
     WELCOME = "Welkom",
     NINJA = "Ninja",
     NOT_YOU = "Ben jij dit niet?",
-    LOGIN_FOR_MORE_FEATURES = "Log in via Github en ontgrendel meer."
+    LOGIN_FOR_MORE_FEATURES = "Log in via Github en ontgrendel meer.",
+    LOGGED_IN_ENCOURAGING = "Ingelogd? Goed bezig."
 }
